+++ migrations/m180103_224855_tickets_messages.php
@@ -17,7 +17,7 @@ class m180103_224855_tickets_messages extends Migration
             $tableOptions = 'CHARACTER SET utf8 COLLATE utf8_unicode_ci ENGINE=InnoDB';
         }
 
-        $this->createTable('{{%tickets_messages%}}', [
+        $this->createTable('{{%tickets_messages}}', [
             'id' => $this->primaryKey(), // Primary key ID (int)
             'ticket_id' => $this->integer()->notNull(), // Ticket ID (int) `tickets`.`id`
             'sender_id' => $this->integer()->null(), // Ticket created by (int) `users`.`id`
@@ -29,7 +29,7 @@ class m180103_224855_tickets_messages extends Migration
 
         $this->createIndex(
             'idx_tickets_messages',
-            '{{%tickets_messages%}}',
+            '{{%tickets_messages}}',
             [
                 'id',
                 'ticket_id',
@@ -37,24 +37,24 @@ class m180103_224855_tickets_messages extends Migration
             ]
         );
 
-        if (!(Yii::$app->db->getTableSchema('{{%tickets%}}', true) === null)) {
+        if (!(Yii::$app->db->getTableSchema('{{%tickets}}', true) === null)) {
             $this->addForeignKey(
                 'fk_messages_to_tickets',
-                '{{%tickets_messages%}}',
+                '{{%tickets_messages}}',
                 'ticket_id',
-                '{{%tickets%}}',
+                '{{%tickets}}',
                 'id',
                 'RESTRICT',
                 'CASCADE'
             );
         }
 
-        if (!(Yii::$app->db->getTableSchema('{{%users%}}', true) === null)) {
+        if (!(Yii::$app->db->getTableSchema('{{%users}}', true) === null)) {
             $this->addForeignKey(
                 'fk_messages_to_users',
-                '{{%tickets_messages%}}',
+                '{{%tickets_messages}}',
                 'sender_id',
-                '{{%users%}}',
+                '{{%users}}',
                 'id',
                 'SET NULL',
                 'CASCADE'
@@ -67,21 +67,21 @@ class m180103_224855_tickets_messages extends Migration
      */
     public function safeDown()
     {
-        if (!(Yii::$app->db->getTableSchema('{{%tickets%}}', true) === null)) {
+        if (!(Yii::$app->db->getTableSchema('{{%tickets}}', true) === null)) {
             $this->dropForeignKey(
                 'fk_messages_to_tickets',
-                '{{%tickets_messages%}}'
+                '{{%tickets_messages}}'
             );
         }
 
-        if (!(Yii::$app->db->getTableSchema('{{%users%}}', true) === null)) {
+        if (!(Yii::$app->db->getTableSchema('{{%users}}', true) === null)) {
             $this->dropForeignKey(
                 'fk_messages_to_users',
-                '{{%tickets_messages%}}'
+                '{{%tickets_messages}}'
             );
         }
 
-        $this->truncateTable('{{%tickets_messages%}}');
-        $this->dropTable('{{%tickets_messages%}}');
+        $this->truncateTable('{{%tickets_messages}}');
+        $this->dropTable('{{%tickets_messages}}');
     }
 }