+++ migrations/m180103_225247_tickets_attachments.php
@@ -17,7 +17,7 @@ class m180103_225247_tickets_attachments extends Migration
             $tableOptions = 'CHARACTER SET utf8 COLLATE utf8_unicode_ci ENGINE=InnoDB';
         }
 
-        $this->createTable('{{%tickets_attachments%}}', [
+        $this->createTable('{{%tickets_attachments}}', [
             'id' => $this->primaryKey(), // Primary key ID (int)
             'ticket_id' => $this->integer()->notNull(), // Ticket ID (int) `tickets`.`id`
             'sender_id' => $this->integer()->null(), // Attachment uploaded by (int) `users`.`id`
@@ -29,7 +29,7 @@ class m180103_225247_tickets_attachments extends Migration
 
         $this->createIndex(
             'idx_tickets_attachments',
-            '{{%tickets_attachments%}}',
+            '{{%tickets_attachments}}',
             [
                 'id',
                 'ticket_id',
@@ -37,36 +37,36 @@ class m180103_225247_tickets_attachments extends Migration
             ]
         );
 
-        if (!(Yii::$app->db->getTableSchema('{{%tickets_messages%}}', true) === null)) {
+        if (!(Yii::$app->db->getTableSchema('{{%tickets_messages}}', true) === null)) {
             $this->addForeignKey(
                 'fk_messages_to_attachments',
-                '{{%tickets_messages%}}',
+                '{{%tickets_messages}}',
                 'attachment_id',
-                '{{%tickets_attachments%}}',
+                '{{%tickets_attachments}}',
                 'id',
                 'SET NULL',
                 'CASCADE'
             );
         }
 
-        if (!(Yii::$app->db->getTableSchema('{{%users%}}', true) === null)) {
+        if (!(Yii::$app->db->getTableSchema('{{%users}}', true) === null)) {
             $this->addForeignKey(
                 'fk_attachments_to_users',
-                '{{%tickets_attachments%}}',
+                '{{%tickets_attachments}}',
                 'sender_id',
-                '{{%users%}}',
+                '{{%users}}',
                 'id',
                 'SET NULL',
                 'CASCADE'
             );
         }
 
-        if (!(Yii::$app->db->getTableSchema('{{%tickets%}}', true) === null)) {
+        if (!(Yii::$app->db->getTableSchema('{{%tickets}}', true) === null)) {
             $this->addForeignKey(
                 'fk_attachments_to_tickets',
-                '{{%tickets_attachments%}}',
+                '{{%tickets_attachments}}',
                 'ticket_id',
-                '{{%tickets%}}',
+                '{{%tickets}}',
                 'id',
                 'RESTRICT',
                 'CASCADE'
@@ -79,28 +79,28 @@ class m180103_225247_tickets_attachments extends Migration
      */
     public function safeDown()
     {
-        if (!(Yii::$app->db->getTableSchema('{{%tickets_messages%}}', true) === null)) {
+        if (!(Yii::$app->db->getTableSchema('{{%tickets_messages}}', true) === null)) {
             $this->dropForeignKey(
                 'fk_messages_to_attachments',
-                '{{%tickets_messages%}}'
+                '{{%tickets_messages}}'
             );
         }
 
-        if (!(Yii::$app->db->getTableSchema('{{%users%}}', true) === null)) {
+        if (!(Yii::$app->db->getTableSchema('{{%users}}', true) === null)) {
             $this->dropForeignKey(
                 'fk_attachments_to_users',
-                '{{%tickets_attachments%}}'
+                '{{%tickets_attachments}}'
             );
         }
 
-        if (!(Yii::$app->db->getTableSchema('{{%tickets%}}', true) === null)) {
+        if (!(Yii::$app->db->getTableSchema('{{%tickets}}', true) === null)) {
             $this->dropForeignKey(
                 'fk_attachments_to_tickets',
-                '{{%tickets_attachments%}}'
+                '{{%tickets_attachments}}'
             );
         }
 
-        $this->truncateTable('{{%tickets_attachments%}}');
-        $this->dropTable('{{%tickets_attachments%}}');
+        $this->truncateTable('{{%tickets_attachments}}');
+        $this->dropTable('{{%tickets_attachments}}');
     }
 }